post-review fixes, moved Identity to /shared
