+++ apps/allocations/app/app-state-reducer.js
@@ -1,10 +1,6 @@
 function appStateReducer(state) {
-  const { github, repos } = state || {}
-  
-  
   return {
     ...state,
-  
   }
 }
