+++ shared/identity/IdentityManager.js
