+++ shared/identity/LocalIdentityBadge.js
