+++ shared/identity/index.js
@@ -1,0 +1,2 @@
+export { IdentityConsumer, IdentityContext, IdentityProvider, useIdentity } from './IdentityManager'
+export { default as LocalIdentityBadge } from './LocalIdentityBadge'
